requirements.txt
arabic-reshaper==3.0.0
asgiref==3.2.10
asn1crypto==1.5.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
colorlog==6.7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
cssselect2==0.7.0
Django==3.1
django-admin-interface==0.20.0
django-colorfield==0.7.2
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-maintenance-mode==0.16.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
drf-yasg2==1.19.4
html5lib==1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
num2words==0.5.12
oscrypto==1.3.0
packaging==21.3
Pillow==9.2.0
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.4.0
pyparsing==3.0.9
pypdf==3.6.0
pypng==0.20220715.0
python-bidi==0.4.2
python-decouple==3.6
python-fsutil==0.6.1
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
svglib==1.5.1
tinycss2==1.2.1
twilio==7.17.0
typing_extensions==4.5.0
tzdata==2022.2
tzlocal==4.3
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.12
webencodings==0.5.1
xhtml2pdf==0.2.9
